samples-python/python.webserver/requirements.txt
flatbuffers~=1.12
autopep8
PyJWT